safari-534.53-branch/LayoutTests/java/lc3/JavaObject/JavaObjectToShort-005-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion JavaObject to short

PASS dt.PUB_DOUBLE_REPRESENTATION = 32767.6;dt.setShort( dt ); dt.PUB_SHORT is 32767
PASS dt.PUB_DOUBLE_REPRESENTATION = 32767.6;dt.setShort( dt ); dt.getShort() is 32767
PASS typeof dt.getShort() is "number"
PASS successfullyParsed is true

TEST COMPLETE
 
